Better heuristic
I was using the squared euclidean distance as an
heuristic, which turns out to not be admisible.
Now the heuristic uses the regural euclidean distance.
I've also tweaked it a bit, to penalize the posible movements
that change of direction, to make the paths prettier.
Also, when moving in 4 directions (when horizontal movement
is off) use manhatan distance instead.